Python/Covid-19 Data Visualization Python/requirements.txt
requests~=2.22.0
pandas~=2.0.3
plotly~=5.15.0
python-dotenv~=1.0.0